// gauvendi-platform-service-v2/src/modules/room-product-feature-rate-adjustment/enums/room-product-feature-rate-adjustment-validation-message.enum.ts
export enum RoomProductFeatureRateAdjustmentValidationMessage {
  // Required field validation
  NULL_INPUT = 'NULL_INPUT',
  REQUIRED_ROOM_PRODUCT_RATE_PLAN_ID = 'REQUIRED_ROOM_PRODUCT_RATE_PLAN_ID',
  REQUIRED_FEATURE_ID = 'REQUIRED_FEATURE_ID',
  REQUIRED_RATE_ADJUSTMENT = 'REQUIRED_RATE_ADJUSTMENT',
  REQUIRED_RATE_ORIGINAL = 'REQUIRED_RATE_ORIGINAL',
  REQUIRED_FROM_DATE = 'REQUIRED_FROM_DATE',
  REQUIRED_TO_DATE = 'REQUIRED_TO_DATE',

  // Success messages
  CREATE_SUCCESS = 'CREATE_SUCCESS',
  UPDATE_SUCCESS = 'UPDATE_SUCCESS',
  DELETE_SUCCESS = 'DELETE_SUCCESS',
  PURGE_SUCCESS = 'PURGE_SUCCESS',
}

export const RoomProductFeatureRateAdjustmentValidationMessages = {
  [RoomProductFeatureRateAdjustmentValidationMessage.NULL_INPUT]: 'PRICING_MODULE-ROOM_PRODUCT_FEATURE_RATE_ADJUSTMENT-400',
  [RoomProductFeatureRateAdjustmentValidationMessage.REQUIRED_ROOM_PRODUCT_RATE_PLAN_ID]: 'PRICING_MODULE-ROOM_PRODUCT_FEATURE_RATE_ADJUSTMENT-401',
  [RoomProductFeatureRateAdjustmentValidationMessage.REQUIRED_FEATURE_ID]: 'PRICING_MODULE-ROOM_PRODUCT_FEATURE_RATE_ADJUSTMENT-402',
  [RoomProductFeatureRateAdjustmentValidationMessage.REQUIRED_RATE_ADJUSTMENT]: 'PRICING_MODULE-ROOM_PRODUCT_FEATURE_RATE_ADJUSTMENT-403',
  [RoomProductFeatureRateAdjustmentValidationMessage.REQUIRED_RATE_ORIGINAL]: 'PRICING_MODULE-ROOM_PRODUCT_FEATURE_RATE_ADJUSTMENT-404',
  [RoomProductFeatureRateAdjustmentValidationMessage.REQUIRED_FROM_DATE]: 'PRICING_MODULE-ROOM_PRODUCT_FEATURE_RATE_ADJUSTMENT-405',
  [RoomProductFeatureRateAdjustmentValidationMessage.REQUIRED_TO_DATE]: 'PRICING_MODULE-ROOM_PRODUCT_FEATURE_RATE_ADJUSTMENT-406',
  [RoomProductFeatureRateAdjustmentValidationMessage.CREATE_SUCCESS]: 'PRICING_MODULE-ROOM_PRODUCT_FEATURE_RATE_ADJUSTMENT-200',
  [RoomProductFeatureRateAdjustmentValidationMessage.UPDATE_SUCCESS]: 'PRICING_MODULE-ROOM_PRODUCT_FEATURE_RATE_ADJUSTMENT-201',
  [RoomProductFeatureRateAdjustmentValidationMessage.DELETE_SUCCESS]: 'PRICING_MODULE-ROOM_PRODUCT_FEATURE_RATE_ADJUSTMENT-202',
  [RoomProductFeatureRateAdjustmentValidationMessage.PURGE_SUCCESS]: 'PRICING_MODULE-ROOM_PRODUCT_FEATURE_RATE_ADJUSTMENT-203',
};